% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAN_llm_prompts.R
\docType{package}
\name{llm_prompts}
\alias{IAN-package}
\alias{llm_prompts}
\title{LLM Prompt Creation Functions}
\description{
This script defines functions to create prompts for Large Language Models (LLMs) based on various enrichment analysis results.
The prompts are designed to guide the LLM in analyzing the data and generating insights about the underlying biological mechanisms.
}
\section{Functions}{

\itemize{
  \item{\code{\link{create_llm_prompt_wp}}}{: Creates an LLM prompt for WikiPathways enrichment analysis.}
  \item{\code{\link{create_llm_prompt_kegg}}}{: Creates an LLM prompt for KEGG enrichment analysis.}
  \item{\code{\link{create_llm_prompt_reactome}}}{: Creates an LLM prompt for Reactome enrichment analysis.}
  \item{\code{\link{create_llm_prompt_chea}}}{: Creates an LLM prompt for ChEA enrichment analysis.}
  \item{\code{\link{create_llm_prompt_go}}}{: Creates an LLM prompt for GO enrichment analysis.}
  \item{\code{\link{create_llm_prompt_string}}}{: Creates an LLM prompt for STRING interaction analysis.}
}
}

